:hammer: refactored pages to new structure
