--- src/pages/bitcoin-data-page/functions/index.ts
@@ -1,3 +1,0 @@
-import { currenciesFormatted } from './currencies-formatted';
-
-export { currenciesFormatted };
