--- src/pages/bitcoin-data-page/styles.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components/native';
-
-import { themes } from '@/themes';
-
-export const PaddingContainer = styled.View`
-	padding: 0 ${themes.spaces.space_15};
-	max-width: ${themes.width.max_width};
-	width: 100%;
-	align-self: center;
-`;
