--- src/pages/local-auth-page/styles.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components/native';
-
-import { themes } from '@/themes';
-
-export const LocalAuthPageContainer = styled.View`
-	flex: 1;
-	background-color: ${themes.colors.black_100};
-	align-items: center;
-	justify-content: center;
-`;
