--- src/pages/register-key-page/styles.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components/native';
-
-import { themes } from '@/themes';
-
-export const TextErrorContainer = styled.View`
-	width: 100%;
-	max-width: ${themes.width.max_width};
-	align-self: center;
-`;
