--- src/pages/transaction-page/types.ts
@@ -1,5 +1,0 @@
-import { AddressTransaction } from '@tatumio/tatum';
-
-export interface TransactionPageProps {
-	data: AddressTransaction;
-}
